Move Cassandra migration related code
All common code is moved to backends-cassandra. It includes:
 - Migration interface: it allows using Cassandra migration utils outside of cassandra-mailbox project
 - Migration service: it allows scheduling cassandra migrations outside of webadmin. It can be a easy move in order to get CLI triggered migration.
