+++ backends-common/cassandra/src/main/java/org/apache/james/backends/cassandra/migration/CassandraMigrationService.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.webadmin.service;
+package org.apache.james.backends.cassandra.migration;
 
 import java.util.Map;
 import java.util.Optional;
@@ -29,9 +29,6 @@ import javax.inject.Named;
 import org.apache.commons.lang.NotImplementedException;
 import org.apache.james.backends.cassandra.versions.CassandraSchemaVersionDAO;
 import org.apache.james.backends.cassandra.versions.CassandraSchemaVersionManager;
-import org.apache.james.mailbox.cassandra.mail.migration.Migration;
-import org.apache.james.mailbox.exception.MailboxException;
-import org.apache.james.webadmin.dto.CassandraVersionResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -53,12 +50,12 @@ public class CassandraMigrationService {
         this.allMigrationClazz = allMigrationClazz;
     }
 
-    public CassandraVersionResponse getCurrentVersion() {
-        return new CassandraVersionResponse(schemaVersionDAO.getCurrentSchemaVersion().join());
+    public Optional<Integer> getCurrentVersion() {
+        return schemaVersionDAO.getCurrentSchemaVersion().join();
     }
 
-    public CassandraVersionResponse getLatestVersion() {
-        return new CassandraVersionResponse(Optional.of(latestVersion));
+    public Optional<Integer> getLatestVersion() {
+        return Optional.of(latestVersion);
     }
 
     public synchronized void upgradeToVersion(int newVersion) {
@@ -76,7 +73,7 @@ public class CassandraMigrationService {
         upgradeToVersion(latestVersion);
     }
 
-    private void doMigration(Integer version) throws MailboxException {
+    private void doMigration(Integer version) {
         int newVersion = version + 1;
         if (allMigrationClazz.containsKey(version)) {
             LOG.info("Migrating to version {} ", newVersion);
